feat: вынести задачи в отдельный компонент
