+++ src/components/InventoryTabs.jsx
@@ -7,8 +7,8 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import { supabase } from '../supabaseClient'
 import { handleSupabaseError } from '../utils/handleSupabaseError'
 import HardwareCard from './HardwareCard'
-import TaskCard from './TaskCard'
 import ChatTab from './ChatTab'
+import TasksTab from './TasksTab'
 import { PlusIcon, ChatBubbleOvalLeftIcon } from '@heroicons/react/24/outline'
 import { linkifyText } from '../utils/linkify'
 import { toast } from 'react-hot-toast'
@@ -16,7 +16,6 @@ import ConfirmModal from './ConfirmModal'
 import Spinner from './Spinner'
 import ErrorMessage from './ErrorMessage'
 import { useHardware } from '../hooks/useHardware'
-import { useTasks } from '../hooks/useTasks'
 import { useChatMessages } from '../hooks/useChatMessages'
 import { useObjects } from '../hooks/useObjects'
 import { useAuth } from '../hooks/useAuth'
@@ -26,19 +25,7 @@ import { apiBaseUrl, isApiConfigured } from '../apiConfig'
 const TAB_KEY = (objectId) => `tab_${objectId}`
 const HW_MODAL_KEY = (objectId) => `hwModal_${objectId}`
 const HW_FORM_KEY = (objectId) => `hwForm_${objectId}`
-const TASK_MODAL_KEY = (objectId) => `taskModal_${objectId}`
-const TASK_FORM_KEY = (objectId) => `taskForm_${objectId}`
 const PAGE_SIZE = 20
-
-// форматирование даты для отображения в русской локали
-function formatDate(dateStr) {
-  if (!dateStr) return ''
-  try {
-    return new Date(dateStr).toLocaleDateString('ru-RU')
-  } catch {
-    return dateStr
-  }
-}
 
 function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
   const navigate = useNavigate()
@@ -90,49 +77,8 @@ function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
   const [hardwareHasMore, setHardwareHasMore] = useState(true)
   const [loadingHW, setLoadingHW] = useState(false)
 
-  // --- задачи ---
-  const [tasks, setTasks] = useState([])
-  const [isTaskModalOpen, setIsTaskModalOpen] = useState(false)
-  const [editingTask, setEditingTask] = useState(null)
-  const defaultTaskForm = {
-    title: '',
-    status: 'запланировано',
-    assignee: '',
-    assignee_id: '',
-    due_date: '',
-    notes: '',
-  }
-  const taskSchema = z.object({
-    title: z.string().min(1, 'Введите название'),
-    status: z.enum(['запланировано', 'в процессе', 'завершено'], {
-      message: 'Выберите статус',
-    }),
-    assignee: z.string().optional(),
-    assignee_id: z.string().optional(),
-    due_date: z.string().optional(),
-    notes: z.string().optional(),
-  })
-  const {
-    register: registerTask,
-    handleSubmit: handleSubmitTask,
-    reset: resetTask,
-    watch: watchTask,
-    formState: { errors: taskErrors },
-  } = useForm({
-    resolver: zodResolver(taskSchema),
-    defaultValues: defaultTaskForm,
-  })
-  const [showDatePicker, setShowDatePicker] = useState(false)
-  const [viewingTask, setViewingTask] = useState(null)
-  const [taskDeleteId, setTaskDeleteId] = useState(null)
-  const taskEffectRan = React.useRef(false)
-  const [tasksError, setTasksError] = useState(null)
-  const [tasksPage, setTasksPage] = useState(0)
-  const [tasksHasMore, setTasksHasMore] = useState(true)
-  const [loadingTasks, setLoadingTasks] = useState(false)
-
   // --- импорт/экспорт ---
-  const [importTable, setImportTable] = useState(null)
+  const [isImportModalOpen, setIsImportModalOpen] = useState(false)
   const [importFile, setImportFile] = useState(null)
 
   // --- чат ---
@@ -143,13 +89,6 @@ function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
     updateHardware,
     deleteHardware,
   } = useHardware()
-  const {
-    fetchTasks: fetchTasksApi,
-    insertTask,
-    updateTask,
-    deleteTask,
-    subscribeToTasks,
-  } = useTasks()
   const { fetchMessages, subscribeToMessages } = useChatMessages()
   const { updateObject } = useObjects()
   // загрузка данных при смене объекта и восстановление состояния UI
@@ -186,47 +125,16 @@ function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
       )
     }
     setIsHWModalOpen(savedHWOpen)
-    const savedTaskForm =
-      typeof localStorage !== 'undefined'
-        ? localStorage.getItem(TASK_FORM_KEY(selected.id))
-        : null
-    const savedTaskOpen =
-      typeof localStorage !== 'undefined'
-        ? localStorage.getItem(TASK_MODAL_KEY(selected.id)) === 'true'
-        : false
-    let parsedTaskForm = defaultTaskForm
-    if (savedTaskForm) {
-      try {
-        parsedTaskForm = JSON.parse(savedTaskForm)
-      } catch {
-        if (typeof localStorage !== 'undefined') {
-          localStorage.removeItem(TASK_FORM_KEY(selected.id))
-        }
-      }
-    }
-    resetTask(parsedTaskForm)
-    if (savedTaskOpen && typeof localStorage !== 'undefined') {
-      localStorage.setItem(
-        TASK_FORM_KEY(selected.id),
-        JSON.stringify(parsedTaskForm),
-      )
-    }
-    setIsTaskModalOpen(savedTaskOpen)
     setDescription(selected.description || '')
 
     setHardware([])
-    setTasks([])
     setChatMessages([])
 
     setHardwarePage(0)
     setHardwareHasMore(true)
     setHardwareError(null)
-    setTasksPage(0)
-    setTasksHasMore(true)
-    setTasksError(null)
 
     fetchHardware(selected.id, 0)
-    fetchTasks(selected.id, 0)
     fetchMessages(selected.id).then(({ data, error }) => {
       if (error) {
         if (error.status === 403) toast.error('Недостаточно прав')
@@ -270,51 +178,9 @@ function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
     return () => sub.unsubscribe()
   }, [watchHW, selected, isHWModalOpen])
 
+  // realtime обновление чата
   useEffect(() => {
     if (!selected) return
-    if (taskEffectRan.current) {
-      if (typeof localStorage !== 'undefined') {
-        localStorage.setItem(TASK_MODAL_KEY(selected.id), isTaskModalOpen)
-        if (!isTaskModalOpen) {
-          localStorage.removeItem(TASK_FORM_KEY(selected.id))
-        }
-      }
-    } else {
-      taskEffectRan.current = true
-    }
-  }, [isTaskModalOpen, selected])
-
-  useEffect(() => {
-    if (!selected || !isTaskModalOpen) return
-    const sub = watchTask((value) => {
-      if (typeof localStorage !== 'undefined') {
-        localStorage.setItem(TASK_FORM_KEY(selected.id), JSON.stringify(value))
-      }
-    })
-    return () => sub.unsubscribe()
-  }, [watchTask, selected, isTaskModalOpen])
-
-  // realtime обновление задач и чата
-  useEffect(() => {
-    if (!selected) return
-    const unsubscribeTasks = subscribeToTasks(selected.id, (payload) => {
-      setTasks((prev) => {
-        if (payload.eventType === 'INSERT') {
-          if (prev.some((t) => t.id === payload.new.id)) return prev
-          return [...prev, payload.new]
-        }
-        if (payload.eventType === 'UPDATE') {
-          return prev.map((t) =>
-            t.id === payload.new.id ? { ...t, ...payload.new } : t,
-          )
-        }
-        if (payload.eventType === 'DELETE') {
-          return prev.filter((t) => t.id !== payload.old.id)
-        }
-        return prev
-      })
-    })
-
     const unsubscribeChat = subscribeToMessages(selected.id, (payload) => {
       setChatMessages((prev) => {
         if (prev.some((m) => m.id === payload.new.id)) return prev
@@ -323,7 +189,6 @@ function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
     })
 
     return () => {
-      unsubscribeTasks()
       unsubscribeChat()
     }
   }, [selected])
@@ -436,121 +301,19 @@ function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
     setHwDeleteId(null)
   }
 
-  // --- CRUD Задачи ---
-
-  async function fetchTasks(objectId, offset = 0) {
-    setLoadingTasks(true)
-
-    setTasksError(null)
-    const { data, error } = await fetchTasksApi(objectId, offset, PAGE_SIZE)
-    if (error) {
-      setTasksError(error)
-      if (error.status === 403) toast.error('Недостаточно прав')
-      else toast.error('Ошибка загрузки задач: ' + error.message)
-    } else {
-      const tasksData = data || []
-      setTasks((prev) => (offset === 0 ? tasksData : [...prev, ...tasksData]))
-      setTasksHasMore(tasksData.length === PAGE_SIZE)
-    }
-    setLoadingTasks(false)
-  }
-
-  function loadMoreTasks() {
-    const nextPage = tasksPage + 1
-    setTasksPage(nextPage)
-    fetchTasks(selected.id, nextPage * PAGE_SIZE)
-  }
-
-  function openTaskModal(item = null) {
-    if (item) {
-      setEditingTask(item)
-      resetTask({
-        title: item.title,
-        status: item.status,
-        assignee: item.assignee || '',
-        assignee_id: item.assignee_id || '',
-        due_date: item.due_date || '',
-        notes: item.notes || '',
-      })
-    } else {
-      setEditingTask(null)
-      resetTask({ ...defaultTaskForm })
-    }
-    setShowDatePicker(false)
-    setIsTaskModalOpen(true)
-  }
-
-  function openTaskView(item) {
-    setViewingTask(item)
-  }
-  async function saveTask(data) {
-    const payload = {
-      object_id: selected.id,
-      title: data.title,
-      status: data.status,
-      assignee: data.assignee || null,
-      assignee_id: data.assignee_id || null,
-      due_date: data.due_date || null,
-      notes: data.notes || null,
-    }
-    let res
-    if (editingTask) {
-      res = await updateTask(editingTask.id, payload)
-    } else {
-      res = await insertTask(payload)
-    }
-
-    if (res.error) {
-      res.error.status === 403
-        ? toast.error('Недостаточно прав')
-        : toast.error('Ошибка задач: ' + res.error.message)
-      await handleSupabaseError(res.error, navigate, 'Ошибка задач')
-      return
-    }
-
-    const rec = res.data
-    setTasks((prev) =>
-      editingTask
-        ? prev.map((t) => (t.id === rec.id ? rec : t))
-        : [...prev, rec],
-    )
-    setIsTaskModalOpen(false)
-    setEditingTask(null)
-    resetTask({ ...defaultTaskForm })
-    toast.success('Задача сохранена')
-  }
-  function askDeleteTask(id) {
-    setTaskDeleteId(id)
-  }
-  async function confirmDeleteTask() {
-    const id = taskDeleteId
-    const { error } = await deleteTask(id)
-
-    if (error) {
-      error.status === 403
-        ? toast.error('Недостаточно прав')
-        : toast.error('Ошибка удаления: ' + error.message)
-      await handleSupabaseError(error, navigate, 'Ошибка удаления')
-      return
-    }
-
-    setTasks((prev) => prev.filter((t) => t.id !== id))
-    setTaskDeleteId(null)
-  }
-
   // --- экспорт/импорт ---
-  function openImportModal(table) {
-    setImportTable(table)
+  function openImportModal() {
+    setIsImportModalOpen(true)
     setImportFile(null)
   }
 
   function closeImportModal() {
-    setImportTable(null)
+    setIsImportModalOpen(false)
     setImportFile(null)
   }
 
   async function handleImport() {
-    if (!importTable || !importFile) return
+    if (!importFile) return
     if (!isApiConfigured) {
       toast.error('API не настроен')
       return
@@ -558,42 +321,35 @@ function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
     const formData = new FormData()
     formData.append('file', importFile)
     try {
-      const res = await fetch(`${apiBaseUrl}/import/${importTable}`, {
+      const res = await fetch(`${apiBaseUrl}/import/hardware`, {
         method: 'POST',
         body: formData,
       })
       if (!res.ok) throw new Error('Ошибка импорта')
       toast.success('Импорт выполнен')
       closeImportModal()
-      if (importTable === 'hardware') {
-        setHardware([])
-        setHardwarePage(0)
-        setHardwareHasMore(true)
-        await fetchHardware(selected.id, 0)
-      } else if (importTable === 'tasks') {
-        setTasks([])
-        setTasksPage(0)
-        setTasksHasMore(true)
-        await fetchTasks(selected.id, 0)
-      }
+      setHardware([])
+      setHardwarePage(0)
+      setHardwareHasMore(true)
+      await fetchHardware(selected.id, 0)
     } catch (e) {
       toast.error(e.message)
     }
   }
 
-  async function handleExport(table, format = 'csv') {
+  async function handleExport(format = 'csv') {
     if (!isApiConfigured) {
       toast.error('API не настроен')
       return
     }
     try {
-      const res = await fetch(`${apiBaseUrl}/export/${table}.${format}`)
+      const res = await fetch(`${apiBaseUrl}/export/hardware.${format}`)
       if (!res.ok) throw new Error('Ошибка экспорта')
       const blob = await res.blob()
       const url = window.URL.createObjectURL(blob)
       const a = document.createElement('a')
       a.href = url
-      a.download = `${table}.${format}`
+      a.download = `hardware.${format}`
       document.body.appendChild(a)
       a.click()
       a.remove()
@@ -623,7 +379,7 @@ function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
           className={`px-4 py-2 hover:bg-primary/10 ${tab === 'tasks' ? 'border-b-2 border-primary' : ''}`}
           onClick={showTasks}
         >
-          Задачи ({tasks.length})
+          Задачи
         </button>
         <button
           className={`px-4 py-2 hover:bg-primary/10 flex items-center gap-1 ${tab === 'chat' ? 'border-b-2 border-primary' : ''}`}
@@ -690,13 +446,13 @@ function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
               <div className="flex gap-2">
                 <button
                   className="btn btn-sm btn-outline"
-                  onClick={() => handleExport('hardware')}
+                  onClick={() => handleExport()}
                 >
                   Экспорт
                 </button>
                 <button
                   className="btn btn-sm btn-outline"
-                  onClick={() => openImportModal('hardware')}
+                  onClick={openImportModal}
                 >
                   Импорт
                 </button>
@@ -850,226 +606,10 @@ function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
 
         {/* Задачи */}
         {tab === 'tasks' && (
-          <div>
-            <div className="flex flex-col sm:flex-row sm:justify-between gap-2 mb-4">
-              <h3 className="text-xl font-semibold">Задачи</h3>
-              <div className="flex gap-2">
-                <button
-                  className="btn btn-sm btn-outline"
-                  onClick={() => handleExport('tasks')}
-                >
-                  Экспорт
-                </button>
-                <button
-                  className="btn btn-sm btn-outline"
-                  onClick={() => openImportModal('tasks')}
-                >
-                  Импорт
-                </button>
-                {user && (
-                  <button
-                    className="btn btn-sm btn-primary flex items-center gap-1"
-                    onClick={() => openTaskModal()}
-                  >
-                    <PlusIcon className="w-4 h-4" /> Добавить задачу
-                  </button>
-                )}
-              </div>
-            </div>
-            {loadingTasks && <Spinner />}
-            {tasksError && (
-              <ErrorMessage
-                error={tasksError}
-                message="Ошибка загрузки задач"
-              />
-            )}
-            {!loadingTasks && !tasksError && (
-              <>
-                {tasks.length === 0 && <p>Задачи не найдены</p>}
-                <div className="grid gap-2 sm:grid-cols-1 md:grid-cols-2 lg:grid-cols-3">
-                  {tasks.map((t) => (
-                    <TaskCard
-                      key={t.id}
-                      item={t}
-                      onView={() => openTaskView(t)}
-                      onEdit={() => openTaskModal(t)}
-                      onDelete={() => askDeleteTask(t.id)}
-                    />
-                  ))}
-                </div>
-              </>
-            )}
-            {tasksHasMore && !loadingTasks && (
-              <button
-                className="btn btn-outline btn-sm mt-2"
-                onClick={loadMoreTasks}
-              >
-                Загрузить ещё
-              </button>
-            )}
-
-            {isTaskModalOpen && (
-              <div className="modal modal-open fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center">
-                <div className="modal-box relative w-full max-w-md p-4 max-h-screen overflow-y-auto animate-fade-in">
-                  <button
-                    className="btn btn-circle absolute right-2 top-2 xs:btn-md md:btn-sm"
-                    onClick={() => setIsTaskModalOpen(false)}
-                  >
-                    ✕
-                  </button>
-                  <h3 className="font-bold text-lg mb-4">
-                    {editingTask ? 'Редактировать' : 'Добавить'} задачу
-                  </h3>
-                  <div className="space-y-4">
-                    <div className="form-control">
-                      <label className="label">
-                        <span className="label-text">Заголовок задачи</span>
-                      </label>
-                      <input
-                        type="text"
-                        className="input input-bordered w-full"
-                        {...registerTask('title')}
-                      />
-                      {taskErrors.title && (
-                        <p className="text-red-500 text-sm mt-1">
-                          {taskErrors.title.message}
-                        </p>
-                      )}
-                    </div>
-                    <div className="form-control">
-                      <label className="label">
-                        <span className="label-text">Исполнитель</span>
-                      </label>
-                      <input
-                        type="text"
-                        className="input input-bordered w-full"
-                        {...registerTask('assignee')}
-                      />
-                      <input type="hidden" {...registerTask('assignee_id')} />
-                      {taskErrors.assignee && (
-                        <p className="text-red-500 text-sm mt-1">
-                          {taskErrors.assignee.message}
-                        </p>
-                      )}
-                    </div>
-                    <div className="form-control">
-                      <label className="label flex items-center">
-                        <span className="label-text">Дата</span>
-                        <button
-                          type="button"
-                          className="ml-2 btn btn-ghost btn-xs"
-                          onClick={() => setShowDatePicker((s) => !s)}
-                        >
-                          📅
-                        </button>
-                      </label>
-                      {showDatePicker && (
-                        <input
-                          type="date"
-                          className="input input-bordered w-full"
-                          {...registerTask('due_date')}
-                        />
-                      )}
-                    </div>
-                    <div className="form-control">
-                      <label className="label">
-                        <span className="label-text">Статус</span>
-                      </label>
-                      <select
-                        className="select select-bordered w-full"
-                        {...registerTask('status')}
-                      >
-                        <option value="запланировано">Запланировано</option>
-                        <option value="в процессе">В процессе</option>
-                        <option value="завершено">Завершено</option>
-                      </select>
-                      {taskErrors.status && (
-                        <p className="text-red-500 text-sm mt-1">
-                          {taskErrors.status.message}
-                        </p>
-                      )}
-                    </div>
-                    <div className="form-control">
-                      <label className="label">
-                        <span className="label-text">Заметки</span>
-                      </label>
-                      <textarea
-                        className="textarea textarea-bordered w-full"
-                        rows={3}
-                        {...registerTask('notes')}
-                      />
-                      {taskErrors.notes && (
-                        <p className="text-red-500 text-sm mt-1">
-                          {taskErrors.notes.message}
-                        </p>
-                      )}
-                    </div>
-                  </div>
-                  <div className="modal-action flex flex-col xs:flex-row space-y-2 xs:space-y-0 xs:space-x-2">
-                    <button
-                      className="btn btn-primary w-full xs:w-auto"
-                      onClick={handleSubmitTask(saveTask)}
-                    >
-                      Сохранить
-                    </button>
-                    <button
-                      className="btn btn-ghost w-full xs:w-auto"
-                      onClick={() => setIsTaskModalOpen(false)}
-                    >
-                      Отмена
-                    </button>
-                  </div>
-                </div>
-              </div>
-            )}
-
-            {viewingTask && (
-              <div className="modal modal-open fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center">
-                <div className="modal-box relative w-full max-w-md p-4 max-h-screen overflow-y-auto animate-fade-in">
-                  <button
-                    className="btn btn-circle absolute right-2 top-2 xs:btn-md md:btn-sm"
-                    onClick={() => setViewingTask(null)}
-                  >
-                    ✕
-                  </button>
-                  <h3 className="font-bold text-lg mb-4">
-                    {viewingTask.title}
-                  </h3>
-                  <div className="space-y-2">
-                    {viewingTask.assignee && (
-                      <p>
-                        <strong>Исполнитель:</strong> {viewingTask.assignee}
-                      </p>
-                    )}
-                    {viewingTask.due_date && (
-                      <p>
-                        <strong>Дата:</strong>{' '}
-                        {formatDate(viewingTask.due_date)}
-                      </p>
-                    )}
-                    <p>
-                      <strong>Статус:</strong> {viewingTask.status}
-                    </p>
-                    {viewingTask.notes && (
-                      <p className="whitespace-pre-wrap break-words">
-                        <strong>Заметки:</strong> {viewingTask.notes}
-                      </p>
-                    )}
-                  </div>
-                </div>
-              </div>
-            )}
-
-            <ConfirmModal
-              open={!!taskDeleteId}
-              title="Удалить задачу?"
-              onConfirm={confirmDeleteTask}
-              onCancel={() => setTaskDeleteId(null)}
-            />
-          </div>
+          <TasksTab selected={selected} onUpdateSelected={onUpdateSelected} />
         )}
 
-        {importTable && (
+        {isImportModalOpen && (
           <div className="modal modal-open fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center">
             <div className="modal-box relative w-full max-w-md p-4 max-h-screen overflow-y-auto animate-fade-in">
               <button
@@ -1078,9 +618,7 @@ function InventoryTabs({ selected, onUpdateSelected, onTabChange = () => {} }) {
               >
                 ✕
               </button>
-              <h3 className="font-bold text-lg mb-4">
-                Импорт {importTable === 'hardware' ? 'оборудования' : 'задач'}
-              </h3>
+              <h3 className="font-bold text-lg mb-4">Импорт оборудования</h3>
               <input
                 type="file"
                 className="file-input file-input-bordered w-full"
